% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges.R
\name{add_edge.multiplex}
\alias{add_edge.multiplex}
\alias{add_edge.mutliplex}
\title{Add an edge between two nodes in a layer}
\usage{
add_edge.multiplex(multiplex, x, y, layer, weight)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{x, y}{: Character, the vertex name attribute}

\item{layer}{: Numeric, the layer's id number where the edge will be added}

\item{weight}{: An integer, the weight of the added edge}
}
\value{
Multiplex network object, whith the edge added
}
\description{
Add an edge between two nodes in a layer
}
\examples{
M <- Multiplex_Lazega
M <- add_edge.multiplex(M,"2","34",2)

}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
