% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges.R
\name{add_edges.multiplex}
\alias{add_edges.multiplex}
\alias{add_edges.mutliplex}
\title{Add a list of edges in a layer}
\usage{
add_edges.multiplex(multiplex, edges, layer)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{edges}{: A list of list, can be a list of tuples(x,y) or tiplets(x,y,w)}

\item{layer}{: Numeric, the layer's id number where the edge will be added}
}
\value{
Multiplex network object, whith the list of edges added
}
\description{
Add a list of edges in a layer of a multiplex network
}
\examples{
M <- Multiplex_Lazega
add_edges.multiplex(M,list(c("1","2"),c("2","3")),3)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
