% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layer.R
\name{add_layer}
\alias{add_layer}
\title{Add an igraph graph as a layer to the multiplex}
\usage{
add_layer(multiplex, g)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{g}{: An igraph graph i.e : the inserted layer. Nodes of the added graph should have a "name" attribute. This attribute is used to match already existing nodes.}
}
\value{
The multiplex graph with the layer added
}
\description{
Add an igraph graph as a layer to the multiplex network. If there is new nodes in the input graph. These new nodes will be added to the multiplex network.
}
\examples{
M <- new("Multiplex")
g1 <- graph.famous("Zachary")
g1 <- set.vertex.attribute(g1,name = "name",index = V(g1),value = as.character(V(g1)))
M <- add_layer(M,g1)
g2 <- graph.ring(30)
g2 <- set.vertex.attribute(g2,name = "name",index = V(g2),value = as.character(V(g2)))
M <- add_layer(M,g2)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
