% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrushCluster.R
\name{add_trushcluster}
\alias{add_trushcluster}
\title{Add a trush cluster}
\usage{
add_trushcluster(memb)
}
\arguments{
\item{memb}{: A membership partition list}
}
\value{
Multiplex network object, whith the edge added
}
\description{
Add a trush cluster
}
\examples{
M <- Multiplex_Lazega

}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
