% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vertex.R
\name{add_vertex.multiplex}
\alias{add_vertex.multiplex}
\title{Adds a new vertex in the multiplex network}
\usage{
add_vertex.multiplex(multiplex,vertex_name)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{vertex_name}{: A string, the vertex name attribute to be added.  This vertex name is not already in the multiplex}
}
\value{
The multiplex network object, with the new node added
}
\description{
Adds a new vertex in the multiplex network
}
\examples{
M <- Multiplex_Lazega
M <- add_vertex.multiplex(M, "72")
# M <- add_vertex.multiplex(M, "1")
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
