% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vertex.R
\name{add_vertices.multiplex}
\alias{add_vertices.multiplex}
\title{Adds a list  of vertex in the multiplex network}
\usage{
add_vertices.multiplex(multiplex,vertices)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{vertices}{: A list of string. It contains the vertices name attribute to be added.  Each vertex name is not already in the multiplex}
}
\value{
The multiplex network object, with the new vertices added
}
\description{
Adds a list of vertex in the multiplex network
}
\examples{
M <- Multiplex_Lazega
add_vertices.multiplex(M, c("72","73","76","145"))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
