% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List2multiplex.R
\name{adjList2multiplex}
\alias{adjList2multiplex}
\title{adjacency matrix list to multiplex}
\usage{
adjList2multiplex(list,nodesName)
}
\arguments{
\item{list}{: A list of adjacency matrix of the same length}

\item{nodesName}{: Vector that contains node's name.}
}
\value{
a multiplex object
}
\description{
Convert a list of adjacency matrix to a multiplex network
}
\examples{
adjMultiplex <- get_adjacency.multiplex(Multiplex_CKM_Physicians_Innovation)
adjList2multiplex(adjMultiplex)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
