% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite2multiplex.R
\name{bipartite2multiplex}
\alias{bipartite2multiplex}
\title{convert a bipartite network to two multiplex network.}
\usage{
bipartite2multiplex(edgeList,subset1,subset2,rankList, threshold1, threshold2, weighted)
}
\arguments{
\item{edgeList}{: The edge list, a two column matrix, character or numeric}

\item{subset1}{: vector. The first subset  (top)}

\item{subset2}{: vector. The second subset (bottom)}

\item{rankList}{: a integer vector. e.g (1:5)}

\item{threshold1}{: numeric in [0,1]. The minimum threshold to consider a link in the first projection}

\item{threshold2}{: numeric in [0,1]. The minimum threshold to consider a link in the second projection}

\item{weighted}{: logical. whether to include the weight generated by the projection.}
}
\value{
a list of two multiplex network
}
\description{
convert a bipartite network to two multiplex network by projecting the first subset on the second subset. And projecting the second subset on the first subset.
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
