% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality_closeness.multiplex}
\alias{centrality_closeness.multiplex}
\title{Centralize a multiplex network according to the multiplex closeness of vertices}
\usage{
centrality_closeness.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The input multiplex network}
}
\value{
A numeric vector with the multiplex  closeness scores.
}
\description{
Centralize a multiplex network according to the closeness of vertices. See closeness.multiplex.
}
\examples{
M <- Multiplex_CS_Aarhus
centrality_closeness.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
