% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality_degree.multiplex}
\alias{centrality_degree.multiplex}
\title{Centralize a multiplex network according to the multiplex degree of vertices}
\usage{
centrality_degree.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The input multiplex network}
}
\value{
A numeric vector with the multiplex degree scores.
}
\description{
s  a  method  for  creating  a  graph  level  centralization  measure  from  the  centrality scores of the vertices. See degree_entropy.multiplex
}
\examples{
M <- Multiplex_CS_Aarhus
centrality_degree.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
