% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community.by_layer}
\alias{community.by_layer}
\title{Community structure detection on each layer}
\usage{
community.by_layer(multiplex,community_algorithm)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{community_algorithm}{: String, the name of classical community detection algorithm. As Licod, Louvain, Infomap, Walktrap,...}
}
\value{
A list of membership of each layer.
}
\description{
Find community structure on each layer separately via classical community detection algorithm. As Licod, Louvain, Infomap, Walktrap,...
}
\examples{
M <- Multiplex_Lazega
community.by_layer(M)
community.by_layer(M,cluster_infomap)
community.by_layer(M,cluster_louvain)
community.by_layer(M,cluster_walktrap)
community.by_layer(M,cluster_edge_betweenness)
community.by_layer(M,cluster_label_prop)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
