% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community.layer_aggregation}
\alias{community.layer_aggregation}
\title{Layer aggregation based community detection algorithm}
\usage{
community.layer_aggregation(multiplex, flatten_method, community_algorithm)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{flatten_method}{: flaten a multiplex network into a simple monoplex network method}

\item{community_algorithm}{: String, the name of classical community detection algorithm. As Licod, Louvain, Infomap, Walktrap,...}
}
\value{
returns an igraph  communities object, please see igraph manual page for details.
}
\description{
This method applies the binary layer aggregation approaches. Then a classical community detection algorithm (one of igraph provided algorithms) can be applied to the resulting monoplex network.
}
\examples{
M <- Multiplex_Lazega
community.layer_aggregation (M, flatten_binary, cluster_louvain)
community.layer_aggregation (M, flatten_redundancy,cluster_louvain)
community.layer_aggregation(M, flatten_binary, cluster_infomap)
community.layer_aggregation(M, flatten_binary, cluster_louvain)
community.layer_aggregation(M, flatten_binary, cluster_walktrap)
community.layer_aggregation(M, flatten_binary, cluster_edge_betweenness)
community.layer_aggregation(M, flatten_binary, cluster_label_prop)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
