% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxlicod.R
\name{community.muxLicod}
\alias{community.muxLicod}
\title{MuxLicod community detection algorithm}
\usage{
community.muxLicod(multiplex,sigma,centrality1,centrality2,delta,vote_fun,sim_fun,memb_fun,verbose)
}
\arguments{
\item{multiplex}{: The input multiplex network}

\item{sigma}{: is a threshold in [0; 1]. It is used to know if a node is a leader.}

\item{centrality1}{: is a topological centrality in graphs as degree, closeness, betweenness...}

\item{centrality2}{: is a topological centrality in graphs as degree, closeness, betweenness...}

\item{delta}{: is a threshold in [0; 1]. Two leaders are linked if their topological similarity is above delta.}

\item{vote_fun}{: The vote function that should be used to compute the membership vector.}

\item{sim_fun}{: is the similarity function. i.e similarity.invlogweighted, similarity.jaccard, similarity.dice,...}

\item{memb_fun}{: is the membership function i.e mean, sum, ...}

\item{verbose}{: verbose}
}
\value{
returns an igraph  communities object, please see igraph manual page for details.
}
\description{
This applies the MuxLicod algorithm as described in EGC 2015 (I.Falih, M.Hmimida, and R.Kanawati)
}
\references{
Issam Falih, Manel Hmimida, Rushed Kanawati, Une approache centr\'ee graine pour la d\'etection de communaut\'es dans les r\'eseaux multiplexes. Conf\'erence francophone sur l'Extraction et la Gestion de Connaissance (EGC'15), 27-30 Jan. 2015 Luxembourg.
#' @examples
M <- Multiplex_CKM_Physicians_Innovation
community.muxLicod(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
