% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity.R
\name{complementarity}
\alias{complementarity}
\title{Complementarity of a multiplex network}
\usage{
complementarity(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A  numeric scalar, the complementarity of the multiplex network
}
\description{
The complementarity  is defined as a conjunction of three measures : variety, homgeniety, exclusivity
}
\examples{
M <- Multiplex_Lazega
complementarity(M)
}
\references{
M. Berlingerio, M. Coscia, and F. Giannotti, Finding and characterizing communities in multidimensional networks, in ASONAM. IEEE Computer Society, 2011, pp. 490-494.
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
