% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity.R
\name{complementarity_partition}
\alias{complementarity_partition}
\title{Partition complementarity of a multiplex network}
\usage{
complementarity_partition(multiplex, partition)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{partition}{: an igraph  communities object, vertex membership.}
}
\value{
A  numeric scalar, the complementarity of the given multiplex network
}
\description{
this is defined as a conjunction of three measures : variety, homgeniety, exclusivity
}
\examples{
M <- Multiplex_Lazega
wt <- community.partition_aggregation(M)
complementarity_partition(M, wt)
}
\references{
M. Berlingerio, M. Coscia, and F. Giannotti, Finding and characterizing communities in multidimensional networks, in ASONAM. IEEE Computer Society, 2011, pp. 490-494.
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
