% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components.multiplex}
\alias{components.multiplex}
\title{The maximal connected components of each layer in the multiplex network}
\usage{
components.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object to analyse.}
}
\value{
A list of the maximal connected components of each layer in the multiplex network. Each components contain a named list with three components:

membership numeric vector giving the cluster id to which each vertex belongs.

csize numeric vector giving the sizes of the clusters.

no  numeric constant, the number of clusters.
}
\description{
Calculate the maximal  connected components of each layer in the multiplex network
}
\examples{
M <- Multiplex_Lazega
components.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
