% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree.multiplex}
\alias{degree.multiplex}
\title{Degree of vertices in the multiplex network}
\usage{
degree.multiplex(multiplex,node_name,delta)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{node_name}{: Character scalar, or list. The vertex name of which the degree vertices are queried.}

\item{delta}{: Numeric scalar, the minimum pourcentage of layers in which the neighbors must be adjacent to the vertex.}
}
\value{
numeric vector of the same length as argument node_name
}
\description{
The degree of vertices in the multiplex network based on the cardinality of the union/intersect neighborhood.
}
\examples{
M <- Multiplex_Lazega
degree.multiplex(M, delta=0.3)
degree.multiplex(M,"2",0.3)
degree.multiplex(M,c("2","5","45"),0.3)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
