% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree_average.multiplex}
\alias{degree_average.multiplex}
\title{Average segree of vertices in the multiplex network}
\usage{
degree_average.multiplex(multiplex,node_name)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{node_name}{: Character scalar, or list. The vertex name of which the degree vertices are queried.}
}
\value{
numeric vector of the same length as argument node_name
}
\description{
The average degree of vertices across all layers in the multiplex network.
}
\examples{
M <- Multiplex_Lazega
degree_average.multiplex(M)
degree_average.multiplex(M,"2")
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
