% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree_entropy.multiplex}
\alias{degree_entropy.multiplex}
\title{Degree based entropy of  vertices in the multiplex network}
\usage{
degree_entropy.multiplex(multiplex,node_name)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{node_name}{: Character scalar, or list. The vertex name of which the degree vertices are queried.}
}
\value{
numeric vector of the same length as argument node_name
}
\description{
The degree of vertices in the multiplex network, the number of its similar neighbors ie neighbors_similarity.multiplex
}
\examples{
M <- Multiplex_Lazega
degree_entropy.multiplex(M)
degree_entropy.multiplex(M,c("2","4","45"))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
