% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree_similarity.multiplex}
\alias{degree_similarity.multiplex}
\title{Degree based similarity of  vertices in the multiplex network}
\usage{
degree_similarity.multiplex(multiplex,node_name,similarity_function,delta)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{node_name}{: Character scalar, or list. The vertex name of which the degree vertices are queried.}

\item{similarity_function}{: Character scalar, the name of the similarity function.}

\item{delta}{: Numeric scalar, the minimum pourcentage of layers in which the neighbors must be adjacent to the vertex.}
}
\value{
Dataframe the name of  vertices and the degree of each vertex.
}
\description{
The degree of vertices in the multiplex network, the number of its similar neighbors ie neighbors_similarity.multiplex
}
\examples{
M <- Multiplex_Lazega
degree_similarity.multiplex(M,"2" ,similarity.jaccard, 0)
degree_similarity.multiplex(M,c("2","4","45"), similarity.jaccard, 0.3)
degree_similarity.multiplex(M, similarity_function = similarity.jaccard, delta=0.8)
degree_similarity.multiplex(M, similarity_function = similarity.invlogweighted,delta= 0.3)
degree_similarity.multiplex(M, similarity_function = similarity.dice, delta=0.3)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
