% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{delete_layer_by_number.multiplex}
\alias{delete_layer_by_number.multiplex}
\title{Delete a layer in the multiplex network}
\usage{
delete_layer_by_number.multiplex(multiplex, id_layer)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{id_layer}{: The id of the layer to delete.}
}
\value{
The multiplex network, with the layer removed
}
\description{
Delete a layer in the multiplex network
}
\examples{
M <- Multiplex_Lazega
M <- delete_layer_by_number.multiplex(M,2)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
