% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecount.R
\name{ecount.multiplex}
\alias{ecount.multiplex}
\title{The number of edges in the multiplex network}
\usage{
ecount.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A list of the number of edges in each layer.
}
\description{
The number of edges in the multiplex network per layer
}
\examples{
M <- Multiplex_Lazega
ecount.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
