% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_threshold_graph.R
\name{epsilon_threshold_graph}
\alias{epsilon_threshold_graph}
\title{Epsilon-threshold graph}
\usage{
epsilon_threshold_graph(mat,epsilon,names,similar)
}
\arguments{
\item{mat}{: distance or similarity matrix. Should be square.}

\item{epsilon}{: numeric in [0,1].}

\item{names}{: character list. The names of vertices.}

\item{similar}{: logical scalar. If true matrix mat is a smiliarity matrix. Otherwise, mat is a distance matrix.}
}
\value{
Returns an igraph graph
}
\description{
Epsilon-threshold graph. Two nodes are connected if their similarity is above
}
\examples{
d <-  dist(iris)
epsilon_threshold_graph(as.matrix(d),epsilon=0.5)

}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
