% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity.R
\name{exclusivity}
\alias{exclusivity}
\title{The exclusivity of the multiplex network}
\usage{
exclusivity(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A  numeric scalar, the exclusivity of the given multiplex network
}
\description{
This is the number of pairs of nodes that are connected exclusively in one layer. The exclusivity is ration of edges in the community belonging to one mono diemensional to the total number of edges in the community.
}
\examples{
M <- Multiplex_CS_Aarhus
exclusivity(M)
}
\references{
M. Berlingerio, M. Coscia, and F. Giannotti, Finding and characterizing communities in multidimensional networks, in ASONAM. IEEE Computer Society, 2011, pp. 490-494.
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
