% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_binary.R
\name{flatten_binary}
\alias{flatten_binary}
\title{Binary layer aggregation}
\usage{
flatten_binary(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A new igraph graph object.
}
\description{
Binary aggregation of all layers in the multiplex network : Edges which are included in at least one graph will be part of the new graph ( the aggregated graph).
}
\examples{
M <- Multiplex_Lazega
g <- flatten_binary(M)
summary_graph(g)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
