% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_MinLayers.R
\name{flatten_MinLayers}
\alias{flatten_MinLayers}
\title{layer aggregation}
\usage{
flatten_MinLayers(multiplex, MinLayers)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{MinLayers}{: minimal number of layer}
}
\value{
A new  igraph graph.
}
\description{
return an igraph graph two nodes are linked in the result graph if they are linked in a least 'MinLayers'  layer.
}
\examples{
M <- Multiplex_Lazega
g <- flatten_MinLayers(M,3)
summary_graph(g)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
