% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_redundancy.R
\name{flatten_redundancy}
\alias{flatten_redundancy}
\title{Redundancy layer aggregation}
\usage{
flatten_redundancy(multiplex, inv=FALSE)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{inv}{: Logical, the weight is inverted if inv is TRUE. By default, inv is FALSE}
}
\value{
A new weighted igraph graph.
}
\description{
return a weighted graph two nodes are linked in the result graph if they are linked i one layer. The edge is weighted by the number of layers in which the two nodes are connected.
}
\examples{
M <- Multiplex_Lazega
g <- flatten_redundancy(M)
summary(g)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
