% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_layer_by_name.multiplex}
\alias{get_layer_by_name.multiplex}
\title{Get a layer in the multiplex network}
\usage{
get_layer_by_name.multiplex(multiplex, name)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{name}{: The name of the layer to be return.}
}
\value{
The igraph graph representing the layer with the name given in parameter, Null if nothing
}
\description{
get a layer in the multiplex network
}
\examples{
library(igraph)
M <- Multiplex_Lazega
g <- get_layer_by_name.multiplex(M, name="Co-work")
str(g)
plot(g)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
