% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_layer_by_number.multiplex}
\alias{get_layer_by_number.multiplex}
\title{Get the multiplex layer which has an id number given in parameters}
\usage{
get_layer_by_number.multiplex(multiplex, layer_number)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{layer_number}{: Numeric scalar, an integer representing the id number of a layer}
}
\value{
An igraph graph representing layer widh id layer_number, NULL if ths id layer_number don't exist
}
\description{
Get the multiplex layer which has an id number given in parameters
}
\examples{
M <- Multiplex_Lazega
g <- get_layer_by_number.multiplex(M,2)
plot(g)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
