% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_layer_name.multiplex}
\alias{get_layer_name.multiplex}
\title{Get the layer's  name attribute of the multiplex network}
\usage{
get_layer_name.multiplex(multiplex, id_layer)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{id_layer}{: The id of the layer in the multiplex network.}
}
\value{
String, the layer's  name attribute.
}
\description{
get the layer's  name attribute of the multiplex network
}
\examples{
M <- Multiplex_Lazega
print(get_layer_name.multiplex(M,3))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
