% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_name.multiplex}
\alias{get_name.multiplex}
\title{Get the attribute name of the multiplex network}
\usage{
get_name.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
String, the name of the multiplex network
}
\description{
get the attribute name of the multiplex network
}
\examples{
M <- Multiplex_Lazega
print(get_name.multiplex(M))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
