% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_number_of_layers.multiplex}
\alias{get_number_of_layers.multiplex}
\title{Get the number of layers in the multiplex network}
\usage{
get_number_of_layers.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
Numeric scalar, an integer representing the number of layers in the multiplex network.
}
\description{
Get the number of layers in the multiplex network
}
\examples{
M <- Multiplex_Lazega
get_number_of_layers.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
