% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity.R
\name{homgeniety}
\alias{homgeniety}
\title{Homgeniety of the multiplex network}
\usage{
homgeniety(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A  numeric scalar, the homgeniety of the given multiplex network
}
\description{
The homgeniety mesures how uniform is the distribution of the number of edgesper layer. The idea that the link intra-community must have a uniform distribution among all layers.
}
\examples{
M <- Multiplex_Lazega
homgeniety(M)
}
\references{
M. Berlingerio, M. Coscia, and F. Giannotti, Finding and characterizing communities in multidimensional networks, in ASONAM. IEEE Computer Society, 2011, pp. 490-494.
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
