% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.multiplex}
\alias{is.multiplex}
\title{Is this object a multiplex network ?}
\usage{
is.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: An R object}
}
\value{
A logical constant, TRUE if argument is a multiplex network
}
\description{
is this object a multiplex network ?
}
\examples{
M <- Multiplex_Lazega
is.multiplex(M)
is.multiplex(numeric(15))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
