% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_graph.R
\name{knn_graph}
\alias{knn_graph}
\title{k-nearest neighbor graph}
\usage{
knn_graph(dx,k,names,similarity)
}
\arguments{
\item{dx}{: matrix that express the similarity or distance between elemenst i and j}

\item{k}{: integer the number of neighbors.}

\item{names}{: the names of vertex in the new graph}

\item{similarity}{: if TRUE then dx is a similarity matrix else it is a distance matrix}
}
\value{
returns the Knn_graph defined over the similarity/Distance matrix dx
}
\description{
k-nearest neighbor graph defined over the similarity or distance matrix
}
\examples{
d <-  dist(iris)
knn_graph(as.matrix(d))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
