% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{modularity_bylayer.multiplex}
\alias{modularity_bylayer.multiplex}
\title{Modularity of a community structure by each layer of a multiplex graph}
\usage{
modularity_bylayer.multiplex(multiplex, partition)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{partition}{: an igraph  communities object, vertex membership.}
}
\value{
A list of numeric scalar. The modularity score over each layer of the given configuration.
}
\description{
Modularity of a community structure by each layer of a multiplex graph
}
\examples{
M <- Multiplex_Lazega
wt <- community.partition_aggregation(M)
modularity_bylayer.multiplex(M, wt)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
