% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{modularity_multiplex}
\alias{modularity_multiplex}
\title{Multiplex modularity of a given community structure of a multiplex network}
\usage{
modularity_multiplex(multiplex, partition)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{partition}{: an igraph  communities object, vertex membership.}
}
\value{
A numeric scalar, the multiplex modularity score of the given configuration
}
\description{
Multiplex modularity of a given community structure of a multiplex network
}
\examples{
M <- Multiplex_Lazega
wt <- community.partition_aggregation(M)
modularity_multiplex(M, wt)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
