% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{multiplex.density}
\alias{multiplex.density}
\title{Graph density of each layer in the multiplex network.}
\usage{
multiplex.density(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A list of real constant. This function returns NaN(=0.0/0.0) for an empty multiplex with zero vertices
}
\description{
The density of a graph is the ratio of the number of edges and the number of possible edges.
}
\examples{
M <- Multiplex_Lazega
multiplex.density(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
