% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{multiplex.summary}
\alias{multiplex.summary}
\title{Short description of the multiplex network}
\usage{
multiplex.summary(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
Character scalar, a short description of the multiplex network. A single sentence
}
\description{
Short summary of the multiplex network
}
\examples{
M <- Multiplex_Lazega
multiplex.summary(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
