% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivity.R
\name{multiplex.transitivity}
\alias{multiplex.transitivity}
\title{Clustering coefficient of each layer in the multiplex network.}
\usage{
multiplex.transitivity(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A list of clustering coefficient of each layer
}
\description{
The clustering coefficient measures the probability that the adjacent vertices of a vertex are connected.
}
\examples{
M <- Multiplex_Lazega
multiplex.transitivity(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
