% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors.multiplex}
\alias{neighbors.multiplex}
\title{Neighboring vertices in the multiplex network}
\usage{
neighbors.multiplex(multiplex,node_name, delta=0.5)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{node_name}{: Character scalar, the vertex name of which the neighbors vertices are queried.}

\item{delta}{: Real scalar, the minimum pourcentage of layers in which the neighbors must be adjacent to the vertex node_name. Should be in [0,1].}
}
\value{
A vertex sequence containing the neighbors of the input vertex
}
\description{
A vertex is a neighbor of another one  in the multiplex network if there are adjacent in delta % layers.
}
\examples{
M <- Multiplex_Lazega
neighbors.multiplex(M,node_name ="2", delta=0.5)
neighbors.multiplex(M,node_name ="2", delta=1)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
