% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankBorda.R
\name{rankBorda}
\alias{rankBorda}
\title{Rank Aggregation via Borda algorithm}
\usage{
rankBorda(mat,decreasing)
}
\arguments{
\item{mat}{: a matrix of ordered lists to be combined (lists must be in rows)}

\item{decreasing}{: logical. Should the sort be increasing or decreasing?}
}
\value{
The aggregated list.
}
\description{
Rank aggregation of ordered lists is performed using the Borda algorithm.
}
\examples{
matt <- matrix(c("A", "B", "C", "D", "E",
"B", "D", "A", "E", "C",
"B", "A", "E", "C", "D",
"A", "D", "B", "C", "E"), byrow=TRUE, ncol=5)
print(rankBorda(matt))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
