% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundency.R
\name{redundency}
\alias{redundency}
\title{The redundency of a multiplex network}
\usage{
redundency(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A  numeric scalar, the redundency of the given multiplex network
}
\description{
The redundency of a multiplex network
}
\examples{
M <- Multiplex_Lazega
redundency(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
