% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundency.R
\name{redundency_partition}
\alias{redundency_partition}
\title{The redundency of each community of a partition in the multiplex network}
\usage{
redundency_partition(multiplex, partition)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{partition}{: an igraph  communities object, vertex membership.}
}
\value{
A list of numeric, the redundency score of each community in a partition.
}
\description{
The redundency of a partition is the average of the redundant links of each intra-community in all layers. The intuition is that the intra-community link should be recurring in different layers.
}
\examples{
M <- Multiplex_Lazega
wt <- community.partition_aggregation(M)
redundency_partition(M, wt)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
