% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parititon_Summary.R
\name{ResultsSummary}
\alias{ResultsSummary}
\title{Results summary}
\usage{
ResultsSummary(multiplex,partition)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{partition}{: an igraph  communities object, vertex membership.}
}
\value{
returns the results of clusters validation indexes
}
\description{
A summary result of a given partition for multiplex
}
\examples{
M <- Multiplex_Lazega
wt <-  community.layer_aggregation(M)
ResultsSummary(M,wt)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
