% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_graph.R
\name{rng_graph}
\alias{rng_graph}
\title{Relative neighbor graph}
\usage{
rng_graph(dx,names,similarity)
}
\arguments{
\item{dx}{: matrix that express the similarity or distance between elements i and j}

\item{names}{: the names of vertex in the new graph}

\item{similarity}{: if TRUE then dx is a similarity matrix else it is a distance matrix}
}
\value{
returns the relative neighborhood graph (RNG) defined over the similarity/Distance matrix dx
}
\description{
the relative neighborhood graph (RNG) is an undirected graph defined on a set of points
         in the Euclidean plane by connecting two points p and q by an edge whenever there does not exist a third point r
                   that is closer to both p and q than they are to each other
}
\examples{
d <-  dist(iris)
rng_graph(as.matrix(d))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
