% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_layer_name.multiplex}
\alias{set_layer_name.multiplex}
\title{Set the layer's name attribute of the multiplex network}
\usage{
set_layer_name.multiplex(multiplex, id_layer, name)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{id_layer}{: The id of the layer in the multiplex network.}

\item{name}{: The new name of the layer.}
}
\value{
The multiplex network, with the layer's name attribute added or set.
}
\description{
set the layer's name attribute of the multiplex network
}
\examples{
M <- Multiplex_Lazega
M <- set_layer_name.multiplex(M, 2, "Co-work layer")
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
