% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_name.multiplex}
\alias{set_name.multiplex}
\title{Set the attribute name of the multiplex network}
\usage{
set_name.multiplex(multiplex, name)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{name}{: The name of the multiplex network.}
}
\value{
The multiplex network, with the name attribute added or set.
}
\description{
set the attribute name of the multiplex network
}
\examples{
M <- Multiplex_Lazega
M <- set_name.multiplex(M, "Lazega Multiplex Network")
print(get_name.multiplex(M))
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr> Rushed Kanawati <rushed.kanawati@lipn.univ-paris13.fr>
}
