% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath.R
\name{shortestPath.multiplex}
\alias{shortestPath.multiplex}
\title{Shortest paths between two vertices in the multiplex network}
\usage{
shortestPath.multiplex(multiplex,nodei,nodej,flatten_function)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{nodei}{: Numeric vector, the vertex from which the shortest paths will be calculated.}

\item{nodej}{: Numeric vector, the vertex to which the shortest paths will be calculated}

\item{flatten_function}{: String, the name of the flatten function}
}
\value{
Numeric scalar
}
\description{
Shortest paths between vertices in the multiplex network. The shortest paths will be calculated in a flatten graph. The shortest path length from a vertex to itself is always zero. For unreachable vertices Inf is included
}
\examples{
M <- Multiplex_Lazega
shortestPath.multiplex(M,"1","34",flatten_function = flatten_binary)
shortestPath.multiplex(M,"1","34",flatten_function = flatten_redundancy)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
