% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath.R
\name{shortestPath_all.multiplex}
\alias{shortestPath_all.multiplex}
\title{Shortest paths between all vertices in the multiplex network}
\usage{
shortestPath_all.multiplex(multiplex,flatten_function)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{flatten_function}{: String, the name of the flatten function}
}
\value{
Numeric square matrix with length of multiplex nodes
}
\description{
Shortest paths between vertices in the multiplex network. The shortest paths will be calculated in a flatten graph. The shortest path length from a vertex to itself is always zero. For unreachable vertices Inf is included
}
\examples{
M <- Multiplex_Lazega
shortestPath_all.multiplex(M,flatten_function = flatten_binary)
shortestPath_all.multiplex(M,flatten_function = flatten_redundancy)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
