% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath.R
\name{shortestPath_mean.multiplex}
\alias{shortestPath_mean.multiplex}
\title{Average shortest paths between vertices over all layers in the multiplex network}
\usage{
shortestPath_mean.multiplex(multiplex,nodei,nodej)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{nodei}{: Numeric vector, the vertex from which the shortest paths will be calculated.}

\item{nodej}{: Numeric vector, the vertex to which the shortest paths will be calculated}
}
\value{
Numeric scalar
}
\description{
Average shortest paths between vertices over all layers in the multiplex network. The shortest path length from a vertex to itself is always zero. For unreachable vertices Inf is included
}
\examples{
M <- Multiplex_Lazega
shortestPath_mean.multiplex(M,"1","34")
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
