% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath.R
\name{shortestPath_mean_all.multiplex}
\alias{shortestPath_mean_all.multiplex}
\title{Average shortest paths between all vertices over all layers in the multiplex network}
\usage{
shortestPath_mean_all.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
Numeric square matrix with length of multiplex nodes
}
\description{
Average shortest paths between all vertices over all layers in the multiplex network. The shortest path length from a vertex to itself is always zero. For unreachable vertices Inf is included
}
\examples{
M <- Multiplex_Lazega
shortestPath_mean_all.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
