% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundency.R
\name{similarity_layer}
\alias{similarity_layer}
\title{The similarity between a input igraph graph and all layers in the multiplex network}
\usage{
similarity_layer(multiplex, graph)
}
\arguments{
\item{multiplex}{: The multiplex object.}

\item{graph}{: an igraph graph}
}
\value{
A  list of numeric scalar,
}
\description{
The similarity between a input igraph graph and all layers in the multiplex network
}
\examples{
M <- Multiplex_Lazega
similarity_layer(M, M@layers[[1]])
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
