% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarityPerlayers}
\alias{similarityPerlayers}
\title{layer similarity of a multiplex network}
\usage{
similarityPerlayers(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
A similarity matrix between layer of the multiplex network
}
\description{
layer similarity of a multiplex network
}
\examples{
M <- Multiplex_Lazega
similarityPerlayers(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
