% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph.R
\name{subgraph.multiplex}
\alias{subgraph.multiplex}
\title{Multiplex subgraph}
\usage{
subgraph.multiplex(multiplex,node_set)
}
\arguments{
\item{multiplex}{: The orginal multiplex object.}

\item{node_set}{: Numeric vector, the vertices of the original multiplex which will form the subgraph.}
}
\value{
A new multiplex object
}
\description{
Creates a multiplex subgraph, containing only the specified vertices and all the edges among them. If the vertex name is not the multiplex network the node will not be considered
}
\examples{
M <- Multiplex_Lazega
SubM <- subgraph.multiplex(M,c("1","2","3"))
multiplex.summary(SubM)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
