% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity.R
\name{variety}
\alias{variety}
\title{Variety of a multiplex network}
\usage{
variety(multiplex)
}
\arguments{
\item{multiplex}{: The multiplex object.}
}
\value{
Numeric scalar, the variety score .
}
\description{
This is the proportion of occurrence of the community  across layers of the multiplex.
}
\examples{
M <- Multiplex_Lazega
variety(M)
}
\references{
M. Berlingerio, M. Coscia, and F. Giannotti, Finding and characterizing communities in multidimensional networks, in ASONAM. IEEE Computer Society, 2011, pp. 490-494.
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
