% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcount.R
\name{vcount.multiplex}
\alias{vcount.multiplex}
\title{The number of vertices in the multiplex network}
\usage{
vcount.multiplex(multiplex)
}
\arguments{
\item{multiplex}{: A multiplex object.}
}
\value{
Numeric scalar, the number of vertices.
}
\description{
The number of vertices in the multiplex network
}
\examples{
M <- Multiplex_Lazega
vcount.multiplex(M)
}
\author{
Issam Falih <issam.falih@lipn.univ-paris13.fr>
}
